requirements.txt
customtkinter==5.2.2
fastapi==0.115.12
uvicorn==0.34.0
firebase-admin==6.8.0
google-auth-oauthlib==1.2.1
google-auth==2.38.0
requests==2.32.3
psutil==7.0.0
pywin32==310
python-dotenv==1.0.1
